+++ cli/src/pcluster/models/cluster.py
@@ -191,6 +191,11 @@ class Cluster:
             self._get_artifact_dir()
         return self.__s3_artifact_dir
 
+    @property
+    def compute_fleet_status_manager(self) -> ComputeFleetStatusManager:
+        """Return compute fleet status manager."""
+        return ComputeFleetStatusManager.get_manager(self.name, self.stack.version, self.stack.scheduler)
+
     def _get_artifact_dir(self):
         """Get artifact directory in S3 bucket by stack output."""
         try:
@@ -272,7 +277,7 @@ class Cluster:
         return status
 
     @property
-    def compute_fleet_status_with_last_updated_time(self):
+    def compute_fleet_status_with_last_updated_time(self) -> Tuple[ComputeFleetStatus, str]:
         """Status of the cluster compute fleet and the last compute fleet status updated time."""
         if self.stack.is_working_status or self.stack.status == "UPDATE_IN_PROGRESS":
             if self.stack.scheduler == "awsbatch":
@@ -281,8 +286,7 @@ class Cluster:
                 )
                 last_updated_time = None
             else:
-                compute_fleet_status_manager = ComputeFleetStatusManager(self.name)
-                status, last_updated_time = compute_fleet_status_manager.get_status_with_last_updated_time()
+                status, last_updated_time = self.compute_fleet_status_manager.get_status_with_last_updated_time()
             return status, last_updated_time
         else:
             LOGGER.info(
@@ -518,7 +522,13 @@ class Cluster:
             if scheduler_plugin_template.startswith("s3"):
                 bucket_parsing_result = parse_bucket_url(scheduler_plugin_template)
                 result = AWSApi.instance().s3.get_object(
-                    bucket_name=bucket_parsing_result["bucket_name"], key=bucket_parsing_result["object_key"]
+                    bucket_name=bucket_parsing_result["bucket_name"],
+                    key=bucket_parsing_result["object_key"],
+                    expected_bucket_owner=get_attr(
+                        self.config,
+                        "scheduling.settings.scheduler_definition.cluster_infrastructure.cloud_formation."
+                        "s3_bucket_owner",
+                    ),
                 )
                 file_content = result["Body"].read().decode("utf-8")
             else:
@@ -530,6 +540,9 @@ class Cluster:
             raise BadRequestClusterActionError(
                 f"Error while downloading scheduler plugin artifacts from '{scheduler_plugin_template}': {str(e)}"
             ) from e
+
+        # checksum
+        self.validate_scheduler_plugin_template_checksum(file_content, scheduler_plugin_template)
 
         # jinja rendering
         try:
@@ -677,15 +690,15 @@ class Cluster:
         except AWSClientError as e:
             raise _cluster_error_mapper(e, f"Failed to retrieve cluster instances. {e}")
 
-    def has_running_capacity(self, updated_value: bool = False):
+    def has_running_capacity(self, updated_value: bool = False) -> bool:
         """Return True if the cluster has running capacity. Note: the value will be cached."""
         if self.__has_running_capacity is None or updated_value:
-            if self.stack.scheduler == "slurm":
-                self.__has_running_capacity = (
-                    ComputeFleetStatusManager(self.name).get_status() != ComputeFleetStatus.STOPPED
-                )
-            elif self.stack.scheduler == "awsbatch":
+            if self.stack.scheduler == "awsbatch":
                 self.__has_running_capacity = self.get_running_capacity() > 0
+            else:
+                self.__has_running_capacity = (
+                    self.compute_fleet_status_manager.get_status() != ComputeFleetStatus.STOPPED
+                )
         return self.__has_running_capacity
 
     def get_running_capacity(self, updated_value: bool = False):
@@ -710,8 +723,8 @@ class Cluster:
             scheduler = self.config.scheduling.scheduler
             if scheduler == "awsbatch":
                 self.enable_awsbatch_compute_environment()
-            else:  # scheduler == "slurm"
-                self.start_slurm_compute_fleet()
+            else:  # traditional scheduler
+                self.start_compute_fleet()
         except ComputeFleetStatusManager.ConditionalStatusUpdateFailed:
             raise BadRequestClusterActionError(
                 "Failed when starting compute fleet due to a concurrent update of the status. "
@@ -720,10 +733,9 @@ class Cluster:
         except Exception as e:
             raise _cluster_error_mapper(e, f"Failed when starting compute fleet with error: {str(e)}")
 
-    def start_slurm_compute_fleet(self):
-        """Start Slurm compute fleet."""
-        compute_fleet_status_manager = ComputeFleetStatusManager(self.name)
-        compute_fleet_status_manager.update_status(
+    def start_compute_fleet(self):
+        """Start compute fleet."""
+        self.compute_fleet_status_manager.update_status(
             ComputeFleetStatus.START_REQUESTED, ComputeFleetStatus.STARTING, ComputeFleetStatus.RUNNING
         )
 
@@ -753,8 +765,8 @@ class Cluster:
             scheduler = self.config.scheduling.scheduler
             if scheduler == "awsbatch":
                 self.disable_awsbatch_compute_environment()
-            else:  # scheduler == "slurm"
-                self.stop_slurm_compute_fleet()
+            else:  # traditional scheduler
+                self.stop_compute_fleet()
         except ComputeFleetStatusManager.ConditionalStatusUpdateFailed:
             raise BadRequestClusterActionError(
                 "Failed when stopping compute fleet due to a concurrent update of the status. "
@@ -763,10 +775,9 @@ class Cluster:
         except Exception as e:
             raise _cluster_error_mapper(e, f"Failed when stopping compute fleet with error: {str(e)}")
 
-    def stop_slurm_compute_fleet(self):
-        """Stop Slurm compute fleet."""
-        compute_fleet_status_manager = ComputeFleetStatusManager(self.name)
-        compute_fleet_status_manager.update_status(
+    def stop_compute_fleet(self):
+        """Stop compute fleet."""
+        self.compute_fleet_status_manager.update_status(
             ComputeFleetStatus.STOP_REQUESTED, ComputeFleetStatus.STOPPING, ComputeFleetStatus.STOPPED
         )
 
@@ -1150,4 +1161,17 @@ class Cluster:
                     "Update failure: The scheduler plugin used for this cluster does not support updating the "
                     "scheduling configuration.",
                     [changes[0]] + scheduling_changes,
+                )
+
+    def validate_scheduler_plugin_template_checksum(self, file_content, scheduler_plugin_template):
+        """Validate scheduler plugin template checksum match the expected checksum."""
+        checksum = get_attr(
+            self.config, "scheduling.settings.scheduler_definition.cluster_infrastructure.cloud_formation.checksum"
+        )
+        if checksum:
+            actual_checksum = hashlib.sha256(file_content.encode()).hexdigest()
+            if actual_checksum != checksum:
+                raise BadRequestClusterActionError(
+                    f"Error when validating scheduler plugin template '{scheduler_plugin_template}': "
+                    f"checksum: {actual_checksum} does not match expected one: {checksum}"
                 )