+++ cli/tests/pcluster/validators/test_scheduler_plugin_validators.py
@@ -11,7 +11,9 @@
 
 import pytest
 
+from pcluster.config.cluster_config import SchedulerPluginUser, SudoerConfiguration
 from pcluster.validators.scheduler_plugin_validators import (
+    GrantSudoPrivilegesValidator,
     SchedulerPluginOsArchitectureValidator,
     SchedulerPluginRegionValidator,
     SudoPrivilegesValidator,
@@ -43,6 +45,40 @@ from tests.pcluster.validators.utils import assert_failure_messages
 )
 def test_sudo_privileges_validator(grant_sudo_privileges, requires_sudo_privileges, expected_message):
     actual_failures = SudoPrivilegesValidator().execute(grant_sudo_privileges, requires_sudo_privileges)
+    assert_failure_messages(actual_failures, expected_message)
+
+
+@pytest.mark.parametrize(
+    "grant_sudo_privileges, system_users, expected_message",
+    [
+        (False, [SchedulerPluginUser(name="name", enable_imds=True)], None),
+        (
+            True,
+            [
+                SchedulerPluginUser(
+                    name="name",
+                    enable_imds=True,
+                    sudoer_configuration=[SudoerConfiguration(commands="ALL", run_as="Root")],
+                )
+            ],
+            None,
+        ),
+        (
+            False,
+            [
+                SchedulerPluginUser(
+                    name="name",
+                    enable_imds=False,
+                    sudoer_configuration=[SudoerConfiguration(commands="ALL", run_as="Root")],
+                )
+            ],
+            "The used scheduler plugin requires the creation of SystemUsers with sudo access. To grant such rights "
+            "please set GrantSudoPrivileges to true.",
+        ),
+    ],
+)
+def test_grant_sudo_privileges_validator(grant_sudo_privileges, system_users, expected_message):
+    actual_failures = GrantSudoPrivilegesValidator().execute(grant_sudo_privileges, system_users)
     assert_failure_messages(actual_failures, expected_message)
 
 
@@ -130,6 +166,17 @@ def test_scheduler_plugin_region_validator(region, supported_regions, expected_m
             "3.0.0b",
             "3.0.1,     3.0.0b",
             None,
+        ),
+        (
+            "3.0.2",
+            ">=3.0.1b1, <=3.1.0",
+            None,
+        ),
+        (
+            "3.0.0a1",
+            ">=3.0.1b1, <=3.1.0",
+            "The installed version 3.0.0a1 is not supported by the scheduler plugin. Supported versions are: "
+            ">=3.0.1b1, <=3.1.0.",
         ),
         (
             "3.1.0",