+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 CHANGELOG
 =========
 
-3.1.0
+x.x.x
+------
+
+**BUG FIXES**
+- Fix cluster stack in `DELETE_FAILED` when deleting a cluster, due to Route53 hosted zone not empty.
+
+3.1.1
 ------
 
 **ENHANCEMENTS**